+++ lib/array/average/method.ts
@@ -1,5 +1,5 @@
 import * as sum from "../sum/method";
 
-const method = (arr: any[], key?: string) => sum(arr, key) / arr.length;
+const method = (arr: any[], path?: string) => sum(arr, path) / arr.length;
 
 export = method;