+++ lib/array/chunk/method.ts
@@ -3,7 +3,7 @@ const method = (arr: any[], size: number): any[][] => {
     {
       length: Math.ceil(arr.length / size),
     },
-    (value: any, index: number) => arr.slice(index * size, index * size + size),
+    (value, index) => arr.slice(index * size, index * size + size),
   );
 };