+++ lib/array/count/method.ts
@@ -1,3 +1,3 @@
-const method = (arr: any[], value: any) => value ? arr.reduce((a, v) => (v === value ? a + 1 : a + 0), 0) : arr.length;
+const method = (arr: any[], value: any) => value ? arr.reduce((a, v) => a + (v === value), 0) : arr.length;
 
 export = method;