+++ lib/array/median/index.ts
@@ -2,18 +2,19 @@ import * as method from "./method";
 
 declare global {
   interface Array<T> {
-    median(key?: string): number;
+    median(path?: string): number;
   }
 }
 
 /**
- * Returns the median value of a given key
+ * Returns the median value of a given path
  * @memberof Array
- * @returns {Array}
+ * @param {string} [path]
+ * @returns {number}
  * @example
  * [1, 1, 2, 4].median(); // 1.5
  * [{foo: 10}, {foo: 10}, {foo: 20}, {foo: 40}].median('foo'); // 15
  */
-Array.prototype.median = function(key) {
-  return method(this, key);
+Array.prototype.median = function(path) {
+  return method(this, path);
 };