+++ lib/array/pluck/method.ts
@@ -1,11 +1,9 @@
-const method = (arr: any[], key: string): any[] => {
-  const keys = key.split(".");
+import { pathToKeys } from "../../utils";
 
-  return arr.map((item) => {
-    keys.map((k) => item = (item && item[k]) || undefined);
+const method = (arr: any[], path: string): any[] => {
+  const keys = pathToKeys(path);
 
-    return item;
-  });
+  return arr.map((item) => keys.reduce((prev, cur) => (prev && prev[cur]) || undefined, item));
 };
 
 export = method;