+++ lib/array/sum/index.ts
@@ -2,20 +2,20 @@ import * as method from "./method";
 
 declare global {
   interface Array<T> {
-    sum(key?: string): number;
+    sum(path?: string): number;
   }
 }
 
 /**
- * Returns the minimum value of a given key
+ * Returns the minimum value of a given path
  * @memberof Array
- * @param {String} [key]
+ * @param {String} [path]
  * @returns {number}
  * @example
  * [1, 2, 3].sum(); // 6
  * [{a: 1}, {a: 2}, {a: 3}].sum('a'); // 6
  * [{a: {b: 1}}, {a: {b: 2}}, {a: {b: 3}}].sum('a.b'); // 6
  */
-Array.prototype.sum = function(key) {
-  return method(this, key);
+Array.prototype.sum = function(path) {
+  return method(this, path);
 };