+++ lib/object/get/method.ts
@@ -1,4 +1,6 @@
-const method = (obj: { [key: string]: any }, key: string): any =>
-  key.split(".").reduce((prev, cur) => prev[cur], obj);
+import { pathToKeys } from "../../utils";
+
+const method = (obj: { [key: string]: any }, key: string): any => pathToKeys(key)
+  .reduce((prev, cur) => prev[cur], obj);
 
 export = method;